Move operator's main function to be a subcommand
This way we can start moving the rest of the commands to be sub-commands
as well.
